% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_package.R
\name{detach_package}
\alias{detach_package}
\title{Detach all versions of a package}
\usage{
detach_package(pkg, character.only = FALSE)
}
\arguments{
\item{pkg}{unquoted name of package you want to detach/unload}

\item{refcols}{A vector of the column names you would like to move to the left side of the dataframe (in order)}
}
\description{
Detach all versions of a package
}
