% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtext.R
\name{fixtext}
\alias{fixtext}
\title{standardize text within columns of a dataframe}
\usage{
fixtext(df, colnames)
}
\arguments{
\item{df}{dataframe whose column text you want to standardize}
}
\description{
standardize text within columns of a dataframe
}
