% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mult.R
\name{import_mult}
\alias{import_mult}
\title{Move a set of columns to the left-hand side of a data frame}
\usage{
import_mult(ft, ...)
}
\arguments{
\item{ft}{A fileTable of files to import and stack}
}
\description{
Move a set of columns to the left-hand side of a data frame
}
