% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_with.R
\name{names_with}
\alias{names_with}
\title{find data frame column names containing specific text}
\usage{
names_with(df, text)
}
\arguments{
\item{df}{dataframe whose colnames you want to check}

\item{text}{text to match on}
}
\description{
find data frame column names containing specific text
}
