% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savexl.R
\name{savexl}
\alias{savexl}
\title{save excel worksheet with simple defaults}
\usage{
savexl(df, dir = dd, filename)
}
\arguments{
\item{df}{dataframe you want to save in a worksheet}

\item{filename}{raw filename without datetime or file extension}
}
\description{
save excel worksheet with simple defaults
}
