% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_sum.R
\name{skim_sum_levels}
\alias{skim_sum_levels}
\title{build an detailed data dictionary with summary statistics}
\usage{
skim_sum_levels(df)
}
\arguments{
\item{df}{dataframe whose columns you would like to summarize}
}
\description{
build an detailed data dictionary with summary statistics
}
\examples{
df <- skim_sum(mtcars)
df
}
