% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortnames.R
\name{sortnames}
\alias{sortnames}
\title{display sorted column names for a dataframe}
\usage{
sortnames(df)
}
\arguments{
\item{df}{dataframe whose colnames you want to see sorted}
}
\description{
display sorted column names for a dataframe
}
